+++ content/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from content.models import *
+from content.models import Query
 
 admin.site.register(Query)
