+++ content/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+
 from profiles.validators import phone_regex
