+++ content/views.py
@@ -2,29 +2,9 @@ from django.contrib.messages.views import SuccessMessageMixin
 from django.shortcuts import render
 from django.views.generic import CreateView
 
-from users.utils import send_mail
-
 from content.forms import QueryForm
-from plasma_for_covid import settings
-from datetime import datetime
-
 from content.decorators import profile_incomplete_check
-
-
-def send_query(request, query):
-    send_mail(
-        "A new query from PlasmaHelp Website",
-        f"""
-Name : {query.name}
-Email : {query.email}
-Contact Number : {query.contact_number}
-DateTime : {datetime.now()} UTC
-Query : {query.query}
-        """,
-        f"{query.email}",
-        [f"{settings.EMAIL_HOST_USER}"],
-        fail_silently=True,
-    )
+from content.utils import send_query
 
 
 @profile_incomplete_check
