+++ profiles/urls.py
@@ -1,9 +1,6 @@
-from django.conf.urls.static import static
-from django.contrib import admin
 from django.urls import path, include
 
-from plasma_for_covid import settings
-from profiles.views import *
+from profiles.views import ProfileView, ProfileEditView, ProfileCreateView
 
 urlpatterns = [
     path("", ProfileView.as_view(), name="profile"),
