+++ users/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
+
 from users.models import User
 
-# Register your models here.
 
 admin.site.register(User)
