+++ users/utils.py
@@ -1,4 +1,5 @@
 import threading
+
 from django.contrib.auth.tokens import PasswordResetTokenGenerator
 from django.template.loader import render_to_string
 from django.utils.encoding import force_bytes
